Update claw.py
Added a '-d' option to allow for debugging. This skips committing to the repository and simply spits out everything it gets from the machines listed in the .conf file. So the user can glance through each host and command to make sure the output is what they expect. Also added logging for normal operation that will keep track of messed up ssh sessions, such as unknown hosts or authentication failures. Set up to email any errors it catches when not run in debug mode. Fixed a bug where it could get stuck reading commands if the brackets weren't separated by a command or empty line. Plan on adding a check for this in checkFormat.